Add NPM, Node, Express, Moment.js
With the help of Michael Auderer (hermbit), we've installed initalized npm, installed Express and Moment.js.
We locally ran Node to show the files in the srv page. We also have two npm intilization. One in the srv folder, and one that will most likely be moved to the /srv/public folder

Tue Mar  6 14:06:11 EST 2018
juxley@protonmail.com