Integrated qbook.club HTML and CSS into Angular
Here I've created a component call nav-bar and add the HTML and CSS. I will have to edit the styles of the navigation and also change it so it stays with the component. I've added the nav-bar component to app.component.html as well. I believe I can also make the navigation be more appendble.