Routing & Additional Components
Here I've added AppRouting so that we can have mutiple pages. I've
create additional components and match that with the NavBarComponent and
FooterComponent. I've added the following components:
 Account
 Club
 Contact
 Home
 Newsletter (which I may rename to BetaSignUp)
 LogIn
 SignUp

Next, I plan on designin the front end to prepare for user sign up, user
sign in, and the contact page to put in data to soon catch. Most
importantly, I'll need to tighten up the style, design, and responsive
behavior. Perhaps doing this with Bootstrap